requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
dataclasses-json==0.5.7
Deprecated==1.2.13
dill==0.3.6
Django==4.0.6
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.10.1
idna==3.4
joblib==1.2.0
keybert==0.6.0
kombu==5.2.4
marshmallow==3.18.0
marshmallow-enum==1.5.1
multiprocess==0.70.14
mypy-extensions==0.4.3
nltk==3.7
numpy==1.23.4
packaging==21.3
pathos==0.3.0
phonenumbers==8.12.52
Pillow==9.2.0
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pydantic==1.10.2
PyExecJS==1.5.1
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.5
PyYAML==6.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.2
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
user-agent==0.1.10
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1

wget~=3.2
python-magic~=0.4.27